refactor: #172 - 알람 Template 클래스를 패키지로 따로 분리 - Template의 추상 메소드에 protected 접근 제어자를 지정하고, 각 구체 클래스에 public 접근 제어자 지정
