+++ src/main/java/freshtrash/freshtrashbackend/service/alarm/CompleteBidAuctionAlarm.java
@@ -2,6 +2,7 @@ package freshtrash.freshtrashbackend.service.alarm;
 
 import freshtrash.freshtrashbackend.entity.Auction;
 import freshtrash.freshtrashbackend.service.AuctionService;
+import freshtrash.freshtrashbackend.service.alarm.template.AuctionAlarmTemplate;
 import freshtrash.freshtrashbackend.service.producer.AuctionPublisher;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
@@ -15,19 +16,19 @@ public class CompleteBidAuctionAlarm extends AuctionAlarmTemplate {
     }
 
     @Override
-    void update(Long auctionId) {
+    public void update(Long auctionId) {
         this.auctionService.closeAuction(auctionId);
     }
 
     @Override
-    void publishEvent(Auction auction, Long bidMemberId) {
+    public void publishEvent(Auction auction, Long bidMemberId) {
         log.debug("판매자에게 낙찰 알림, 구매자에게 결제 요청 알림");
         this.producer.completeBid(auction, bidMemberId);
         this.producer.requestPay(auction, bidMemberId);
     }
 
     @Override
-    void publishEvent(Auction auction) {
+    public void publishEvent(Auction auction) {
         log.debug("입찰자가 없음을 판매자에게 알림");
         this.producer.notCompleteBid(auction);
     }
