+++ src/main/java/freshtrash/freshtrashbackend/service/alarm/UserFlagChatAlarm.java
@@ -1,6 +1,7 @@
 package freshtrash.freshtrashbackend.service.alarm;
 
 import freshtrash.freshtrashbackend.service.MemberService;
+import freshtrash.freshtrashbackend.service.alarm.template.ChatAlarmTemplate;
 import freshtrash.freshtrashbackend.service.producer.ChatProducer;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
@@ -22,13 +23,13 @@ public class UserFlagChatAlarm extends ChatAlarmTemplate {
      * 유저 신고 횟수 + 1
      */
     @Override
-    int update(Long targetMemberId) {
+    public int update(Long targetMemberId) {
         log.debug("유저 신고 횟수 + 1 업데이트...");
         return this.memberService.updateFlagCount(targetMemberId, FLAG_LIMIT).flagCount();
     }
 
     @Override
-    void publishEvent(int flagCount, Long productId, Long targetMemberId, Long currentMemberId) {
+    public void publishEvent(int flagCount, Long productId, Long targetMemberId, Long currentMemberId) {
         log.debug("현재 신고 횟수 {}", flagCount);
         String message = generateMessage(flagCount);
         this.producer.occurredUserFlag(productId, targetMemberId, currentMemberId, message);
